+++ dubbo-samples-direct/src/main/java/org/apache/dubbo/samples/direct/DirectConsumer.java
@@ -17,20 +17,35 @@
  *
  */
 
-package com.alibaba.dubbo.samples.direct;
+package org.apache.dubbo.samples.direct;
+
+import org.apache.dubbo.samples.direct.api.DirectService;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 
-public class DirectProvicer {
+public class DirectConsumer {
 
-    public static void main(String[] args) throws Exception {
+    public static void main(String[] args) {
         //Prevent to get IPV6 address,this way only work in debug mode
         //But you can pass use -Djava.net.preferIPv4Stack=true,then it work well whether in debug mode or not
         System.setProperty("java.net.preferIPv4Stack", "true");
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/dubbo-direct-provider.xml"});
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/dubbo-direct-consumer.xml"});
         context.start();
+        DirectService directService = (DirectService) context.getBean("directService"); // get remote service proxy
 
-        System.in.read(); // press any key to exit
+        while (true) {
+            try {
+                Thread.sleep(1000);
+                String hello = directService.sayHello("world"); // call remote method
+                System.out.println(hello); // get result
+
+            } catch (Throwable throwable) {
+                throwable.printStackTrace();
+            }
+
+
+        }
+
     }
 }
