+++ dubbo-samples-direct/src/main/java/org/apache/dubbo/samples/direct/DirectProvicer.java
@@ -17,35 +17,20 @@
  *
  */
 
-package com.alibaba.dubbo.samples.direct;
-
-import com.alibaba.dubbo.samples.direct.api.DirectService;
+package org.apache.dubbo.samples.direct;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 
-public class DirectConsumer {
+public class DirectProvicer {
 
-    public static void main(String[] args) {
+    public static void main(String[] args) throws Exception {
         //Prevent to get IPV6 address,this way only work in debug mode
         //But you can pass use -Djava.net.preferIPv4Stack=true,then it work well whether in debug mode or not
         System.setProperty("java.net.preferIPv4Stack", "true");
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/dubbo-direct-consumer.xml"});
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/dubbo-direct-provider.xml"});
         context.start();
-        DirectService directService = (DirectService) context.getBean("directService"); // get remote service proxy
 
-        while (true) {
-            try {
-                Thread.sleep(1000);
-                String hello = directService.sayHello("world"); // call remote method
-                System.out.println(hello); // get result
-
-            } catch (Throwable throwable) {
-                throwable.printStackTrace();
-            }
-
-
-        }
-
+        System.in.read(); // press any key to exit
     }
 }
